ProiectCloud/mysite/requirements.txt
social-auth-app-django
python-twitter
azure-storage-file
textblob
urlextract